feat: refactor directories for calendars
