+++ src/components/RangeCalendar/index.ts
@@ -1,0 +1,1 @@
+export * from './RangeCalendar';
